refactor: rename monty_plugin.dart → monty_extension.dart
File name now matches the class it contains (MontyExtension).
All imports updated. No symbol changes.

dart analyze --fatal-infos: No issues found.
